Moved status logic to actions py
