+++ backend/moonstream/api.py
@@ -7,10 +7,11 @@ import time
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
+from . import actions
 from . import data
+from .middleware import MoonstreamHTTPException
 from .routes.address_info import app as addressinfo_api
 from .routes.nft import app as nft_api
-from .routes.status import app as status_api
 from .routes.subscriptions import app as subscriptions_api
 from .routes.streams import app as streams_api
 from .routes.txinfo import app as txinfo_api
@@ -47,9 +48,29 @@ async def now_handler() -> data.NowResponse:
     return data.NowResponse(epoch_time=time.time())
 
 
+@app.get("/status", response_model=data.StatusResponse)
+async def status_handler() -> data.StatusResponse:
+    """
+    Get latest records and their creation timestamp for crawlers:
+    - ethereum_txpool
+    - ethereum_trending
+    """
+    try:
+        crawl_types_timestamp = actions.check_api_status()
+    except actions.StatusAPIException:
+        raise MoonstreamHTTPException(status_code=500)
+    except Exception as e:
+        logger.error(f"Unhandled status exception, error: {e}")
+        raise MoonstreamHTTPException(status_code=500)
+
+    return data.StatusResponse(
+        ethereum_txpool_timestamp=crawl_types_timestamp["ethereum_txpool"],
+        ethereum_trending_timestamp=crawl_types_timestamp["ethereum_trending"],
+    )
+
+
 app.mount("/subscriptions", subscriptions_api)
 app.mount("/users", users_api)
-app.mount("/status", status_api)
 app.mount("/streams", streams_api)
 app.mount("/txinfo", txinfo_api)
 app.mount("/address_info", addressinfo_api)
